Split folder browser tag editor view
