+++ src/gui_app/folder_browser.rs
@@ -321,6 +321,8 @@ pub(super) use types::{
     FolderScanProgress, FolderScanRequest, FolderScanResult, RemovedSource, RenameTargetView,
 };
 
+mod tag_completion;
+mod tag_editor;
 mod view;
 pub(super) use view::folder_browser_view;
